New organisation of the code.
