+++ Hotel.cs
@@ -1,0 +1,24 @@
+using System;
+using System.Collections.Generic;
+namespace HotelDisplay;
+class Hotel
+{
+
+    public string name;
+    public string location;
+    public string Name{
+        get{return name;}
+        set{name = value;}
+    }
+    public string Location{
+        get{return location;}
+        set{location = value;}
+    }
+
+    public Hotel(string name , string location)
+    {
+        Name = name;
+        Location = location;
+    }
+
+}
